bot/requirements.txt
aiogram<4.0
python-dotenv==1.0.0